requirements.txt
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1
djangorestframework==3.9.0
httpie==1.0.2
idna==2.8
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
